Update main.py
Exception Handling: Exception handling has been added to catch possible exceptions and log error messages. This way, even if an exception occurs in an asynchronous task, the code will continue to execute without causing the entire task to terminate.

Check for Complete NMEA Message Sentences: When reading data from the serial port, a check has been added to see if the data received is a complete NMEA message sentence. If the received data is not complete, it will be ignored until a complete sentence is received.